Добавил ограничение на обработку задач, пока может только админ, пото…

+++ app/Http/Controllers/Tasks/TaskController.php
@@ -54,6 +54,8 @@ class TaskController extends Controller
 
                             ]);
 
+        $this->authorize('viewAny', Task::class);
+
         $data = $request->validated();
 
         $filter = app()->make(TaskHistoryFilter::class, ['queryParams' => array_filter($data)]);
@@ -90,6 +92,8 @@ class TaskController extends Controller
                 'user' => Auth::user()->name,
             ]);
 
+        $this->authorize('create', Task::class);
+
         $users = User::where('superior_uuid', 'like', Auth::id())->orWhere('id', 'like', Auth::id())->get();
 
         $documents = Document::orderBy('created_at', 'desc')->get();
@@ -113,6 +117,8 @@ class TaskController extends Controller
                 'user' => Auth::user()->name,
                 'task' => $task->id,
             ]);
+
+        $this->authorize('create', Task::class);
 
         $users = User::where('superior_uuid', 'like', Auth::id())->orWhere('id', 'like', Auth::id())->get();
 
@@ -139,6 +145,9 @@ class TaskController extends Controller
                 'user' => Auth::user()->name,
                 'request' => $request->all(),
             ]);
+
+        $this->authorize('create', Task::class);
+
 
         if($request->isMethod('post')) {
 
@@ -196,6 +205,9 @@ class TaskController extends Controller
                 'task' => $task->id,
             ]);
 
+        $this->authorize('view', Task::class);
+
+
         return view('tasks.show', [
             'task' => $task,
         ]);
@@ -215,6 +227,8 @@ class TaskController extends Controller
                 'task' => $task->id,
 
             ]);
+
+        $this->authorize('update', Task::class);
 
         return view('tasks.edit', [
             'task' => $task,
@@ -239,6 +253,8 @@ class TaskController extends Controller
                 'task' => $task->id,
                 'request' => $request->all(),
             ]);
+
+        $this->authorize('update', Task::class);
 
         if($request->isMethod('patch')) {
 
@@ -302,6 +318,8 @@ class TaskController extends Controller
                 'task' => $task->id,
             ]);
 
+        $this->authorize('delete', Task::class);
+
         $task->delete();
         return redirect()->route('tasks.index');
     }
@@ -314,6 +332,8 @@ class TaskController extends Controller
                 'task' => $task->id,
                 'document' => $document->id,
             ]);
+
+        $this->authorize('delete', Task::class);
 
         $task_file = TaskFile::where('task_uuid', $task->id)->where('file_uuid', $document->id)->delete();
 
@@ -334,6 +354,8 @@ class TaskController extends Controller
                 'task' => $task->id,
 
             ]);
+
+        $this->authorize('view', Task::class);
 
         return view('tasks.progress', [
             'task' => $task,
@@ -358,6 +380,8 @@ class TaskController extends Controller
                 'task' => $task->id,
                 'request' => $request->all(),
             ]);
+
+        $this->authorize('update', Task::class);
 
         if($request->isMethod('patch')) {
